Add ground truth test Move self test to a separate file. Change to bin count to `uint32_t` due to potential of integer overflow when radius is very large.

+++ Program/DisRegRep/Format.hpp
@@ -14,7 +14,7 @@ using SizeVec2 = std::array<size_t, 2u>;/**< An array of two size_t's. */
 using Radius_t = std::uint16_t;/**< Type of radius. */
 
 using Region_t = std::uint8_t;/**< Type of region, which is also the pixel format of region map. */
-using Bin_t = std::uint16_t;/**< Format of bin that stores region count. */
+using Bin_t = std::uint32_t;/**< Format of bin that stores region count. */
 using NormBin_t = float;/**< Format of bin that stores normalised weight. */
 
 }
